@@ -27,70 +27,69 @@
     "demo:release": "yarn demo:build && node scripts/gh-pages"
   },
   "dependencies": {
-    "@angular/animations": "^7.2.14",
-    "@angular/cdk": "^7.3.7",
-    "@angular/common": "^7.2.14",
-    "@angular/compiler": "^7.2.14",
-    "@angular/core": "^7.2.14",
-    "@angular/forms": "^7.2.14",
-    "@angular/http": "^7.2.14",
-    "@angular/platform-browser": "^7.2.14",
-    "@angular/platform-browser-dynamic": "^7.2.14",
-    "@angular/router": "^7.2.14",
+    "@angular/animations": "^8.2.12",
+    "@angular/cdk": "^8.2.3",
+    "@angular/common": "^8.2.12",
+    "@angular/compiler": "^8.2.12",
+    "@angular/core": "^8.2.12",
+    "@angular/forms": "^8.2.12",
+    "@angular/platform-browser": "^8.2.12",
+    "@angular/platform-browser-dynamic": "^8.2.12",
+    "@angular/router": "^8.2.12",
     "classlist.js": "^1.1.20150312",
     "core-js": "^2.6.3",
-    "date-fns": "^2.0.0-alpha.27",
-    "rxjs": "~6.5.1",
+    "date-fns": "^2.6.0",
+    "rxjs": "~6.5.3",
     "semver": "^5.6.0",
-    "zone.js": "~0.8.29"
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.13.0",
-    "@angular-devkit/build-ng-packagr": "^0.13.0",
-    "@angular/cli": "~7.3.8",
-    "@angular/compiler-cli": "^7.2.14",
-    "@angular/language-service": "^7.2.14",
+    "@angular-devkit/build-angular": "~0.803.14",
+    "@angular-devkit/build-ng-packagr": "~0.803.14",
+    "@angular/cli": "~8.3.14",
+    "@angular/compiler-cli": "^8.2.12",
+    "@angular/language-service": "^8.2.12",
     "@salesforce-ux/design-system": "^2.8.3",
     "@stackblitz/sdk": "^1.3.0",
     "@types/jasmine": "^2.8.9",
     "@types/jasminewd2": "^2.0.6",
     "@types/node": "~8.9.4",
     "app-root-path": "^2.2.0",
-    "codelyzer": "^4.5.0",
-    "conventional-changelog-cli": "^2.0.12",
+    "codelyzer": "^5.0.1",
+    "conventional-changelog-cli": "^2.0.25",
     "del": "^3.0.0",
-    "gh-pages": "^2.0.1",
-    "gulp": "^4.0.0",
+    "gh-pages": "^2.1.1",
+    "gulp": "^4.0.2",
     "gulp-changed": "^3.2.0",
     "gulp-data": "^1.3.1",
     "gulp-pug": "^4.0.1",
     "inquirer": "^6.2.2",
     "jasmine-core": "~2.99.1",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~3.0.0",
+    "karma": "~4.0.0",
     "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "^2.0.5",
+    "karma-coverage-istanbul-reporter": "~2.0.1",
     "karma-firefox-launcher": "^1.1.0",
     "karma-ie-launcher": "^1.0.0",
     "karma-jasmine": "~1.1.2",
     "karma-jasmine-html-reporter": "^0.2.2",
     "karma-sauce-launcher": "^2.0.2",
     "markdown-it": "^8.4.2",
-    "ng-packagr": "^4.7.1",
+    "ng-packagr": "^5.4.0",
     "npm-run-all": "^4.1.5",
     "prismjs": "^1.15.0",
     "protractor": "^5.4.2",
-    "pug": "^2.0.3",
+    "pug": "^2.0.4",
     "q": "^1.5.1",
-    "replace": "^1.1.0",
+    "replace": "^1.1.1",
     "rimraf": "^2.6.3",
     "shelljs": "^0.8.3",
     "simple-git": "^1.107.0",
     "ts-node": "~7.0.0",
-    "tsickle": "^0.34.3",
-    "tslib": "^1.9.0",
+    "tsickle": "^0.37.0",
+    "tslib": "^1.10.0",
     "tslint": "^5.14.0",
-    "typescript": "~3.2.4"
+    "typescript": "~3.5.3"
   },
   "publishConfig": {
     "registry": "https://registry.npmjs.org/"
