@@ -19,13 +19,13 @@
   "author": "Tasos Bekos <tbekos@gmail.com>",
   "license": "MIT",
   "dependencies": {
-    "date-fns": "^2.0.0-alpha.27"
+    "date-fns": "^2.6.0"
   },
   "peerDependencies": {
-    "@angular/cdk": "^7.0.0",
-    "@angular/common": "^7.0.0",
-    "@angular/core": "^7.0.0",
-    "@angular/forms": "^7.0.0"
+    "@angular/cdk": "^8.0.0",
+    "@angular/common": "^8.0.0",
+    "@angular/core": "^8.0.0",
+    "@angular/forms": "^8.0.0"
   },
   "publishConfig": {
     "registry": "https://registry.npmjs.org/"
